LayoutTests/vibration/navigator-vibration.html

<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test for the Vibration API.');

shouldBeTrue("navigator.vibrate(0);");
shouldBeTrue("navigator.vibrate([]);");
shouldBeTrue("navigator.vibrate(1);");
shouldBeTrue("navigator.vibrate([1, 2]);");
shouldBeTrue("navigator.vibrate([1, 2, 3]);");
shouldThrow("navigator.vibrate();");
shouldThrow("navigator.vibrate([1], [2]);");
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
